requirements.txt
playwright
pandas
streamlit
numpy
requests

flask
flask-sqlalchemy
psycopg2-binary
flask-migrate
flask-limiter

python-dotenv
flask-cors

gunicorn

